Code Update #2
Stores each uploaded file and file name in a persistent data structure–in this case, a text file. I may change this to a database later if I have time, however, I think that a text file does an equally good job of storing the file names as it allows for quick searching and updating as the user makes more POST API calls.