Code Update #3
Generate unique fileIDs for each file; return said fileID upon the POST API call. For instance, an ouptut would be:

  "You uploaded:": {
    "file name": "hello.txt"
  }, 
  "fileID:": 1

Here, the fileID is 1 since this is the first entry in the uploaded.txt file. The system for generating fileIDs is quite simple. The line number on which the file name is situated is the same number assigned to the fileID. Therefore, if I were to invoke the POST API call once again, this time on a different file, here would be a sample output:

  "You uploaded:": {
    "file name": "test.txt"
  }, 
  "fileID:": 2

As you can see, the fileID is now 2 since it is the second entry in the uploaded.txt file. Since this is a persistent data structure, we can stop running the server, re-run it, and invoke yet another POST API call and, assuming the file name is different, notice that the user will receive a unique fileID associated with the file they just uploaded.

My next step is to add functionality such that a user does not encounter filename collisions, thus always ensuring that the user receives a unique fileID.