Code Update #4
Fully created functions that handle filename collisions (for instance: if hello.txt is already in the list of uploaded files and the user invokes the POST API with hello.txt as the file they want zipped, the uploaded list will enter this as hello_1.txt. Similarly, if a different user also invokes the POST API with hello.txt, the uploaded list will enter this as hello_2.txt.

Additionally, docstrings were created and updated for all current functions.