Code Update #9 [FINAL update + reflections]
Update: Added assertion tests in Python for zip, ind, file_size (and thus, byteConvert), generateID, and duplicateName (and thus, changeBase). Also cleaned code, added comments/docstrings. Attempted to create aliases for the curl commands (for ease of uploading and retrieving files), however, the shell script would have to be re-inputted every time the program is run due to the nature of Repl.it. Also attempted to upload files to Google Drive (using pyDrive and Google API) instead of storing them locally in a folder, but once again, the IDE was very limiting and I learned that after progressing through a majority of the project. It is also a reason as to why I was unable to run unit tests in a separate .py file and had to resort to assertion tests.

Reflections: The reason I was unable to use a normal IDE like Atom/VSCode/Sublime was due to my python modules not importing correctly, perhaps due to an issue with the pip installer or my copy of python itself–did not have the time to debug this, so I looked for a work-around. In conclusion, this opportunity gave me great insights into the world of software engineering and the numerous roadblocks I faced led to more and more learning opportunities. I want to thank the recruiting team for giving me the opportunity to interview and complete this assignment; I have definitely acquired a new skillset. This experience has truly shaped my perspective on SWE, and for the better! :)